sm8250-common: fstab: Omit defaults from fs_mgr flags
AES-256-XTS with AES-256-CTS for filename encryption is the default on AOSP
and can be completely omitted, as long as the colons are still there.

Encryption policy v2 is the also default for API 30 and newer, launching
devices, but we don't remove it from Virtual A/B fstab to keep uniformity.

Test: None
Link: https://source.android.com/security/encryption/file-based#enabling-fbe-on-internal-storage
Link: https://source.android.com/security/encryption/metadata#switching-on-metadata-encryption
Signed-off-by: Albert I <kras@raphielgang.org>